some udates

+++ app/views.py
@@ -1,3 +1,6 @@
+import json
+from django.template import RequestContext
+from django.http import JsonResponse
 from django.http import HttpResponse
 from django.shortcuts import redirect
 from .models import Board, Card, BoardList
@@ -6,6 +9,7 @@ from django.http import HttpResponseRedirect
 from django.views.generic import TemplateView
 from .forms import BoardForm, ListForm, CardForm, UserForm, UserSignupForm
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, get_object_or_404
 from django.contrib import messages
 
@@ -18,36 +22,39 @@ class Boards(TemplateView):
 
     template_name = 'app/index.html'
 
-    def get(self, request):
-        # select current user and activation
-        boards = Board.objects.filter(user=request.user, activation=True)
-        return render(request, self.template_name, {'boards' : boards})
+    def get(self, *args, **kwargs):
+        if not self.request.user.is_authenticated:
+            return redirect('login')
+        boards = Board.objects.filter(user=self.request.user, activation=True)
+        return render(self.request, self.template_name, {'boards' : boards})
 
 
 class AddBoard(TemplateView):
     """
     Let the user add new board
     """
-
     form = BoardForm
     template_name = 'app/create_board.html'
-    def get(self,request):
+    def get(self, *args, **kwargs):
+        if not self.request.user.is_authenticated:
+            return redirect('login')
         form = self.form()
-        return render(request, self.template_name, {'forms': form})
+        return render(self.request, self.template_name, {'forms': form})
 
-    def post(self, request):
+    def post(self, *args, **kwargs):
         # create a form instance and populate it with data from the request:
-        form = self.form(request.POST)
+        form = self.form(self.request.POST)
         # check whether it's valid:
         if form.is_valid():
+            # import pdb; pdb.set_trace()
             # if it is valid, save(commit=False) gets you a model object, then you can add your extra data
-            form = form.save(commit=False)
+            board = form.save(commit=False)
             # add user data then save() it
-            form.user = request.user
-            form.save()
+            board.user = self.request.user
+            board.save()
             # redirect to a home url:
-            return HttpResponseRedirect('home')
-        return render(request, self.template_name, {'forms': form})
+            return redirect('board_view', board.id)
+        return render(self.request, self.template_name, {'forms': form})
 
 
 class BoardView(TemplateView):
@@ -56,8 +63,9 @@ class BoardView(TemplateView):
     """
 
     template_name = 'app/board_view.html'
-    
     def get(self, *args, **kwargs):
+        if not self.request.user.is_authenticated:
+            return redirect('login')
         board_id = kwargs.get('id')
         # iterate board model, if data doesnt exist it will move you to 404 page
         board = get_object_or_404(Board, id=board_id)
@@ -75,6 +83,8 @@ class AddList(TemplateView):
 
     template_name = 'app/create_list.html'
     def get(self, *args, **kwargs):
+        if not self.request.user.is_authenticated:
+            return redirect('login')
         lists = self.lists()
         return render(self.request, self.template_name, {'lists': lists})
 
@@ -103,6 +113,8 @@ class AddCard(TemplateView):
     template_name_post = 'board_view'
 
     def get(self, *args, **kwargs):
+        if not self.request.user.is_authenticated:
+            return redirect('login')
         form = self.form()
         return render(self.request, self.template_name, {'forms' : form})
 
@@ -130,10 +142,12 @@ class EditCard(TemplateView):
     template_name = 'app/card_edit.html'
 
     def get(self, *args, **kwargs):
+        if not self.request.user.is_authenticated:
+            return redirect('login')
         card_id = kwargs.get('id')
         card = get_object_or_404(Card, id=card_id)
         form = self.form(instance = card)
-        return render(self.request, self.template_name, {'forms' : form})
+        return render(self.request, self.template_name, {'forms' : form, 'card' : card})
 
     def post(self, *args, **kwargs):
         card_id = kwargs.get('id')
@@ -143,7 +157,7 @@ class EditCard(TemplateView):
             form.save()
             return redirect('board_view', card.boardList.board.id)
         else:
-            return render(self.request, self.template_name, {'forms' : form})
+            return render(self.request, self.template_name, {'forms' : form, 'card': card})
 
 
 class DeleteList(TemplateView):
@@ -153,6 +167,8 @@ class DeleteList(TemplateView):
 
     template_name = 'board_view'
     def get(self, request, *args, **kwargs):
+        if not self.request.user.is_authenticated:
+            return redirect('login')
         list_id = kwargs.get('id')
         _list = get_object_or_404(BoardList, id=list_id)
         _list.delete()
@@ -167,6 +183,8 @@ class DeleteCard(TemplateView):
     template_name = 'board_view'
 
     def get(self, request, *args, **kwargs):
+        if not self.request.user.is_authenticated:
+            return redirect('login')
         card_id = kwargs.get('id')
         card = get_object_or_404(Card, id=card_id)
         card.delete()
@@ -251,6 +269,7 @@ class LogoutView(TemplateView):
 
     def get(self, *args, **kwargs):
         logout(self.request)
+        return redirect('home')
 
 
 class EditList(TemplateView):
@@ -262,6 +281,8 @@ class EditList(TemplateView):
     template_name = 'app/list_edit.html'
 
     def get(self, *args, **kwargs):
+        if not self.request.user.is_authenticated:
+            return redirect('login')
         list_id = kwargs.get('id')
         list_object = get_object_or_404(BoardList, id=list_id)
         form = self.form(instance = list_object)
@@ -276,3 +297,140 @@ class EditList(TemplateView):
             return redirect('board_view', list_object.board.id)
         else:
             return render(self.request, self.template_name, {'forms' : form})
+
+
+class ShowCards(TemplateView):
+    """
+    Show Cards
+    """
+    
+    template_name = 'app/show_cards.html'
+    def get(self, *args, **kwargs):
+        # if not self.request.user.is_authenticated:
+        #     return redirect('login')
+        card_id = kwargs.get('id')
+        cards = Card.objects.filter(id=card_id)
+        return render(self.request, self.template_name, {'cards' : cards})
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+class SampleInput(TemplateView):
+    """
+    for dev only :D
+    """
+
+    form = CardForm
+    template_name = 'app/dev.html'
+
+    def get(self, *args, **kwargs):
+        form = self.form()
+        return render(self.request, self.template_name, {'forms' : form})
+
+
+    def post(self, *args, **kwargs):
+        if self.request.method == "POST":
+            boardList = BoardList.objects.get(id='98')
+            form = self.form(self.request.POST)
+            if(form.is_valid()):
+                form = form.save(commit=False)
+                form.boardList = boardList
+                form.save()
+            return HttpResponse(json.dumps({'forms' : form.title}))
+        return render_to_response('app/dev.html',{'form':self.form()}, RequestContext(self.request))
+
+
+class OutputAjax(TemplateView):
+    """
+    dev test
+    """
+
+    template_name = 'app/output_ajax.html'
+    def get(self, *args, **kwargs):
+        cards = Card.objects.filter(boardList=98)
+        return render(self.request, self.template_name, {'cards' : cards})
+    
+    # def post(self, *args, **kwargs):
+
+class OutputAjaxs(TemplateView):
+    """
+    dev test
+    """
+
+    template_name = 'app/dev.html'
+    def get(self, *args, **kwargs):
+        list_id = kwargs.get('id')
+        cards = Card.objects.filter(id=list_id)
+        # import pdb; pdb.set_trace()
+        return render(self.request, self.template_name, {'cards' : cards})
